requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# LangChain and LLM integrations
langchain==0.1.6
langchain-google-genai==0.0.9
langchain-community==0.0.20
langchain-text-splitters==0.0.1

# Google Generative AI
google-generativeai==0.3.2

# Vector store and embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# PDF processing
pypdf==3.17.4

# Data handling
pydantic==2.5.3
